+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinUtil.kt
@@ -1,22 +1,23 @@
 package com.bitmovin.analytics.bitmovin.player
 
 import com.bitmovin.analytics.utils.Util
-import com.bitmovin.player.BuildConfig
 import com.bitmovin.player.api.Player
 
 internal object BitmovinUtil {
-    val playerVersion: String
-        get() {
-            try {
-                val versionField = BuildConfig::class.java.getField("VERSION_NAME")
-                return versionField[null] as String
-            } catch (e: NoSuchFieldException) {
-            } catch (e: IllegalAccessException) {
-            }
-            return "unknown"
-        }
+    val playerVersion: String by lazy {
+        getVersion("com.bitmovin.player.BuildConfig")
+            // Workaround Bitmovin player 3.34 and 3.35
+            ?: getVersion("com.bitmovin.player.core.BuildConfig")
+            ?: "unknown"
+    }
 
     fun getCurrentTimeInMs(player: Player): Long {
         return Util.secondsToMillis(player.currentTime)
     }
+
+    private fun getVersion(buildConfigName: String): String? = kotlin.runCatching {
+        Class.forName(buildConfigName, /* initialize = */ true, Player::class.java.classLoader)
+            .getField("VERSION_NAME")
+            .get(null) as String?
+    }.getOrNull()
 }