+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinUtilTest.kt
@@ -1,0 +1,12 @@
+package com.bitmovin.analytics.bitmovin.player
+
+import com.bitmovin.player.BuildConfig
+import org.junit.Assert
+import org.junit.Test
+
+class BitmovinUtilTest {
+    @Test
+    fun getPlayerVersion_returnsValueFromPlayerBuildConfig() {
+        Assert.assertEquals(BuildConfig.VERSION_NAME, BitmovinUtil.playerVersion)
+    }
+}